refactor: util omfg
