+++ desktopexporter/internal/frontend/src/components/TraceDetails/DetailView/DetailView.svelte
@@ -1,16 +1,52 @@
+<script module lang="ts">
+  import type { AttributeScope, FieldDefinition } from '@/constants/fields'
+
+  /** When empty, show all detail rows; otherwise only matching search fields / attributes. */
+  export function detailSearchFieldVisible(
+    selected: FieldDefinition[],
+    searchFieldName: string
+  ): boolean {
+    if (selected.length === 0) return true;
+    return selected.some(
+      f =>
+        f.searchScope === 'field' &&
+        'name' in f &&
+        f.name === searchFieldName
+    );
+  }
+
+  export function detailAttributeVisible(
+    selected: FieldDefinition[],
+    key: string,
+    attributeScope: AttributeScope
+  ): boolean {
+    if (selected.length === 0) return true;
+    return selected.some(
+      f =>
+        f.searchScope === 'attribute' &&
+        'name' in f &&
+        'attributeScope' in f &&
+        f.name === key &&
+        f.attributeScope === attributeScope
+    );
+  }
+
+  /** Duration is not a search field; tie visibility to start/end time columns. */
+  export function detailDurationVisible(selected: FieldDefinition[]): boolean {
+    if (selected.length === 0) return true;
+    return (
+      detailSearchFieldVisible(selected, 'startTime') ||
+      detailSearchFieldVisible(selected, 'endTime')
+    );
+  }
+</script>
+
 <script lang="ts">
   import type { SpanData } from '@/types/api-types'
-  import type { FieldDefinition } from '@/constants/fields'
-  import {
-    detailSearchFieldVisible,
-    detailAttributeVisible,
-    detailDurationVisible,
-  } from '@/utils/detail-column-filter'
   import SpanField from './SpanField.svelte'
   import EventsPanel from './EventsPanel.svelte'
   import LinksPanel from './LinksPanel.svelte'
-  import { formatDuration } from '@/utils/duration'
-  import { formatTimestamp } from '@/utils/time'
+  import { formatDuration, formatTimestamp } from '@/utils/time'
   import { getTimeContext } from '@/contexts/time-context.svelte'
 
   type Props = {
